requirements.txt
annotated-types==0.7.0
anyio==4.8.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
mypy==1.14.1
mypy-extensions==1.0.0
pydantic==2.10.5
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
uvicorn==0.34.0
selenium==4.22.0
webdriver-manager==3.8.6
bs4
pandas
pandas-stubs

appnope==0.1.4
asttokens==3.0.0
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
emoji==2.14.1
executing==2.2.0
fonttools==4.59.0
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.0
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
statsmodels==0.14.5
textblob==0.19.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13

sqlalchemy
pytest
pymysql-1.1.1
pymongo-4.14.0